# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/escalade-manuelle/{tacheId}": {
      "post": {
        "operationId": "AdminController_escaladerManuellement",
        "parameters": [
          {
            "name": "tacheId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "nomClient": {
                      "type": "string"
                    },
                    "montantTtc": {
                      "type": "number"
                    },
                    "roleCorbeille": {
                      "type": "string"
                    },
                    "ordre": {
                      "type": "number"
                    },
                    "typeActeur": {
                      "type": "string",
                      "enum": [
                        "V",
                        "A",
                        "C"
                      ]
                    },
                    "bloquant": {
                      "type": "boolean"
                    },
                    "slaHeures": {
                      "type": "number"
                    },
                    "modeAffectation": {
                      "type": "string",
                      "enum": [
                        "PULL"
                      ]
                    },
                    "etat": {
                      "type": "string",
                      "enum": [
                        "EN_ATTENTE",
                        "EN_CORBEILLE",
                        "RECLAMEE",
                        "APPROUVEE",
                        "REJETEE",
                        "POST_CLOTURE"
                      ]
                    },
                    "agentClaimId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "dateClaim": {
                      "type": "string",
                      "nullable": true
                    },
                    "verrouExpireAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "echeanceSla": {
                      "type": "string",
                      "nullable": true
                    },
                    "niveauEscalade": {
                      "type": "number"
                    },
                    "dateDecision": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "reference",
                    "nomClient",
                    "montantTtc",
                    "roleCorbeille",
                    "ordre",
                    "typeActeur",
                    "bloquant",
                    "slaHeures",
                    "modeAffectation",
                    "etat",
                    "agentClaimId",
                    "dateClaim",
                    "verrouExpireAt",
                    "echeanceSla",
                    "niveauEscalade",
                    "dateDecision"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/audit/securite": {
      "get": {
        "operationId": "AuditController_journalSecurite",
        "parameters": [
          {
            "name": "utilisateur",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "evenement",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOGIN",
                "LOGOUT",
                "MFA_CHALLENGE",
                "RBAC_REFUS",
                "SOD_REFUS"
              ]
            }
          },
          {
            "name": "depuis",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jusqua",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "default": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/{demandeId}": {
      "get": {
        "operationId": "AuditController_journalDemande",
        "parameters": [
          {
            "name": "demandeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/{demandeId}/export": {
      "get": {
        "operationId": "AuditController_exporter",
        "parameters": [
          {
            "name": "demandeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "pdf"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "PGD — API",
    "description": "Plateforme de Gestion des Dégrèvements — Orange Côte d'Ivoire",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}